Move `g_cancellable_set_error_if_cancelled()` to manual
Regression from this change:
https://github.com/gtk-rs/gtk-rs-core/commit/7c9b36c270a945012a2f5d07c8d0f11912d66d34#diff-8439ea6875c9b60b7d07a7b39f6dde1d31d38459f7c27173e56b4cecb0eadefaR142

Basically `g_cancellable_set_error_if_cancelled()` doesn't follow the
GError conventions - it's the inverse, it returns `TRUE` if an error
was set, and `FALSE` otherwise.

We need to move this to manual implementation.

I caught this while updating ostree-rs-ext to use gio 0.15,
and our https://github.com/ostreedev/ostree-rs-ext/blob/main/lib/src/tokio_util.rs
unit test started to panic.
