+++ gio/Gir.toml
@@ -398,6 +398,9 @@ manual_traits = ["CancellableExtManual"]
     name = "cancelled"
     # Racy, use 'connect' and 'disconnect' instead
     ignore = true
+    [[object.function]]
+    name = "set_error_if_cancelled"
+    manual = true
 
 [[object]]
 name = "Gio.CharsetConverter"
